% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bam.R
\name{bamtolist}
\alias{bamtolist}
\title{Convert BAM files into a list of data tables or into a GRangesList object.}
\usage{
bamtolist(bamfolder, annotation, transcript_align = TRUE, list_name = NULL,
  rm_version = FALSE, granges = FALSE)
}
\arguments{
\item{bamfolder}{A character string indicating the path to the folder
containing the BAM files.}

\item{annotation}{A data table from \code{\link{create_annotation}}. Please
make sure that the name of the reference sequences in the annotation data
table coincides with those in the BAM files.}

\item{transcript_align}{A logical value whether or not the BAM files within
\code{bamfolder} refers to a transcriptome alignment (intended as an
alignment based on a reference FASTA of all the transcript sequences). When
this parameter is TRUE (the default) no reads mapping on the negative
strand should be present and they are therefore removed.}

\item{list_name}{A character string vector specifying the desired names for
the data tables of the output list. Its length must coincides with the
number of BAM files within \code{bamfolder}. Please pay attention to the
order in which they are provided: the first string is assigned to the first
file, the second string to the second one and so on. By default this
argument is NULL, implying that the data tables are named after the name of
the BAM file, leaving their path and extension out.}

\item{rm_version}{A logical value whether ot not to remove the version of the
transcripts from the end of their ID, usually separated by a dot. This
option might be useful to make the transcripts IDs in the BAM files match
with those in the annotation table. Default is FALSE.}

\item{granges}{A logical value whether or not to return a GRangesList object.
Default is FALSE, meaning that a list of data tables (the required input
for \code{\link{length_filter}}, \code{\link{psite}} and
\code{\link{psite_info}}, \code{\link{rends_heat}} and
\code{\link{rlength_distr}}) is returned instead.}
}
\value{
A list of data tables or a GRangesList object.
}
\description{
Reads one or several BAM files, converts each file into a data table and
combines them into a list. Alternatively, it returns a GRangesList i.e. a
list of GRanges objects. In both cases the data structure contains for each
read the name of the reference sequence (i.e. of the transcript) on which it
aligns, the leftmost and rightmost position of the read and its length. Two
additional columns are attached, reporting the leftmost and rightmost
position of the CDS of the reference sequence with respect to its 1st
nuclotide. Please note: if a transcript is not associated to any annotated
CDS then its start and the stop codon are set to 0.
}
\examples{
## path_bam <- "path/to/BAM/files"
## annotation_dt <- datatable_with_transcript_annotation
## bamtolist(bamfolder = path_bam, annotation = annotation_dt)
}
