% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_coverage.R
\name{codon_coverage}
\alias{codon_coverage}
\title{Number of reads per codon.}
\usage{
codon_coverage(
  data,
  annotation,
  sample = NULL,
  psite = FALSE,
  min_overlap = 1,
  output_class = "datatable"
)
}
\arguments{
\item{data}{Either list of data tables or GRangesList object from
\code{\link{psite_info}}. Data tables and GRanges objects generated by
\code{\link{bamtolist}} and \code{\link{bedtolist}} can be used if
\code{psite} is FALSE (the default).}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Character string vector specifying the name of the sample(s) of
interest. Default is NULL i.e. all samples in \code{data} are processed.}

\item{psite}{Logical value whether to return the number of P-sites per codon.
Default is TRUE. If FALSE, the number of read footprints per codon is
returned instead.}

\item{min_overlap}{Positive integer specifying the minimum number of
overlapping positions (in nucleotides) between reads and codons to be
considered overlapping. If \code{psite} is TRUE this parameter must be 1
(the default).}

\item{output_class}{Either "datatable" or "granges". It specifies the format
of the output i.e. a list of data tables or a GRangesList object. Default
is "datatable".}
}
\value{
A data table or GRanges object.
}
\description{
This function computes transcript-specific codon coverages, defined as the
number of either read footprints or P-sites mapping on each triplet of coding
sequences and UTRs (see \emph{Details}). The resulting data table contains,
for each triplet: i) the name of the corresponding reference sequence (i.e.
of the transcript to which it belongs); ii) its leftmost and rightmost
position with respect to the 1st nucleotide of the reference sequence; iii)
its position with respect to the 1st and the last codon of the annotated CDS
of the reference sequence; iv) the region of the transcript (5' UTR, CDS, 3'
UTR) it is in; v) the number of read footprints or P-sites falling in that
region for all samples.
}
\details{
The sequence of every transcript is divided in triplets starting
  from the annotated translation initiation site (if any) and proceeding
  towards the UTRs extremities, possibly discarding the exceeding 1 or 2
  nucleotides at the extremities of the transcript. Please note: transcripts
  not associated to any annotated 5' UTR, CDS and 3'UTR and transcripts whose
  coding sequence length is not divisible by 3 are automatically discarded.
}
\examples{
## data(reads_list)
## data(mm81cdna)
##
## ## compute and add p-site datails
## psite_offset <- psite(reads_list, flanking = 6, extremity = "auto")
## reads_psite_list <- psite_info(reads_list, psite_offset)
##
## Compute the codon coverage based on the number of ribosome footprint per
## codon, setting the minimum overlap between reads and triplets to 3 nts:
## coverage_dt <- codon_coverage(reads_psite_list, mm81cdna, min_overlap = 3)
## 
## Compute the coverage based on the number of P-sites per codon:
## coverage_dt <- codon_coverage(reads_psite_list, mm81cdna, psite = TRUE)
}
