% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaheatmap_psite}
\alias{metaheatmap_psite}
\title{Plot ribosome occupancy metaheatmaps at single-nucleotide resolution.}
\usage{
metaheatmap_psite(data, annotation, sample, scale_factors = NULL,
  length_range = "all", transcripts = NULL, utr5l = 25, cdsl = 50,
  utr3l = 25, log = F, colour = "black", plot_title = NULL)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{annotation}{A data frame as generated by
\code{\link{create_annotation}}.}

\item{sample}{A list of character string vectors specifying the name of the
samples (or of its replicates) of interest. The elements of each vector are
merge together using the scale factors specified by \code{scale_factors}.
The name of the elements of the list are used for labelling the raws of the
heatmap.}

\item{scale_factors}{A numeric vector of scale factors for merging the
replicates (if any). The vector must contain at least a set of values, one
for each replicates, named after the strings listed in \code{sample}. No
specific order is required. Default is NULL, meaning that all the scale
factors are set to 1.}

\item{length_range}{Either "all", an integer or an integer vector. Default is
"all", meaning that all the read lengths are included in the analysis.
Otherwise, only the read lengths matching the specified value(s) are kept.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the analysis. By default this argument is
NULL, meaning that all the transcripts in \code{data} are used. Note that
if either the 5' UTR, the coding sequence or the 3' UTR of a transcript is
shorther than \code{utr5l}, \eqn{2*}\code{cdsl} and \code{utr3l}
respectively, the transcript is automatically discarded.}

\item{utr5l}{A positive integer specifying the length (in nucleotides) of the
5' UTR region that in the plot flanks the start codon. The default value is
25.}

\item{cdsl}{A positive integer specifying the length (in nucleotides) of the
CDS region that in the plot will flank both the start and the stop codon.
The default value is 50.}

\item{utr3l}{A positive integer specifying the length (in nucleotides) of the
3' UTR region that in the plot flanks the start codon. The default value is
25.}

\item{log}{A logical value whether or not to use a logarithmic scale colour
(strongly suggested in case of large variations of the signal). Default is
FALSE.}

\item{colour}{A character string specifying the colour of the plot. Default
is "black".}

\item{plot_title}{Either "auto", NULL or any character string. The default is
"auto", meaning that the title of the plot is the name of the sample
followed by the number of the transcripts and the read lengths considered
for generating the metaprofile.}
}
\value{
A list containing a ggplot2 object, a data frame with the associated
  data and the transcripts employed for generating the plot.
}
\description{
For one or more sample this function plots a heatmap-like metaprofile based
on the P-site of the reads mapping around the start and the stop codon of the
annotated CDS (if any). It works similarly to \code{\link{metaprofile_psite}}
but the intensity of the signal is represented by a continuous color scale
rather than by the height of a line chart. This graphical output is a good
option for analyzing several samples at once and for comparing the profiles
generated by different reads lengths or in multiple conditions.
}
\examples{
data(reads_psite_list)

## Generate the metaheatmap for all the read lengths
metaheat_whole <- metaheatmap_psite(reads_psite_list, mm81cdna, sample = list("Whole"=c("Samp1")))
metaheat_whole[["plot"]]

## Generate the metaheatmap employing reads of 27, 28 and 29 nucleotides and
a subset of transcripts (for example with at least one P-site mapping on the
translation initiation site)
sample_name <- "Samp1"
sub_reads_psite_list <- subset(reads_psite_list[[sample_name]], psite_from_start == 0)
transcript_names <- as.character(sub_reads_psite_list$transcript)
metaheat_sub <- metaheatmap_psite(reads_psite_list, mm81cdna, sample = list("sub"=sample_name),
length_range = 27:29, transcripts = transcript_names, plot_title = "auto")
metaheat_sub[["plot"]]

## Generate two metaheatmaps displaied in the same plot. In this exampe one
data frame includes all the read lengths while in the other one contains only
reads of 28 nucleotides
sample_name <- "Samp1"
metaheat_df <- list()
metaheat_df[["subsample_28nt"]] <- subset(reads_psite_list[[sample_name]], length == 28)
metaheat_df[["whole_sample"]] <- reads_psite_list[[sample_name]]
names_list <- list("Only_28" = c("subsample_28nt"), "All" = c("whole_sample"))
metaheat_comparison <- metaheatmap_psite(metaheat_df, mm81cdna, sample = names_list)
metaheat_comparison[["plot"]]
}
