% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psite_per_cds.R
\name{psite_per_cds}
\alias{psite_per_cds}
\title{Compute the number of P-sites per coding sequence.}
\usage{
psite_per_cds(data, annotation, start_nts = 0, stop_nts = 0)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{annotation}{A data frame from \code{\link{create_annotation}}.}

\item{start_nts}{A positive integer specifying the number of nucleotides at 
the beginning of the coding sequences to be exluded from the count. Default
is 0.}

\item{stop_nts}{A positive integer specifying the number of nucleotides at 
the end of the coding sequences to be exluded from the count. Default is 0.}
}
\value{
A list of data frames.
}
\description{
Computes the number of P-sites in frame locatd within the coding sequence of 
each transcript for all the samples of the input list. It is possible to 
restrict the analysis to a subsequence of the coding region by exluding a 
chosen number of nucleotides at the beginiing and at the end of the CDS. Only
the transcripts associated to an annotated CDS are considered. The resulting
data frame report the name of the transript along with the length of the
considered region (in nucleotide) and the number of p-sites.
}
\examples{
data(reads_psite_list)
data(mm81cdna)

## Compute the number of p-sites in frame along the whole coding sequence.
psite_cds_list <- psite_per_cds(reads_psite_list, mm81cdna)

## Compute the number of p-sites in frame along the coding sequence exluding
the first 15 nucleotides and the last 10 nucleotides.
psite_cds_list <- psite_per_cds(reads_psite_list, mm81cdna, start_nts = 15, stop_nts = 10)
}
