% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_list}
\alias{reads_list}
\title{Reads information}
\format{A list of data tables with 1 object (named \emph{Samp1}) of 393,338
  rows and 6 variables (the lengths and the distances are expressed in
  nucleotides):
\describe{
  \item{transcript}{Name of the transcript (ENST ID and version, dot
  separated)}
  \item{end5}{Position of the 5' end of the read with respect to the first
  nuclotide of the transcript}
  \item{end3}{Position of the 3' end of the read with respect to the first
  nuclotide of the transcript}
  \item{length}{Length of the read}
  \item{start_pos}{Leftmost position of the annotated CDS with respect to the
  first nuclotide of the transcript}
  \item{stop_pos}{Rightmost position of the annotated CDS with respect to the
  first nuclotide of the transcript}
}}
\usage{
reads_list
}
\description{
This dataset contains details on mapping reads from BAM or BED files.
}
\keyword{datasets}
