% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_psite_list}
\alias{reads_psite_list}
\title{P-sites and reads information}
\format{A list of data tables with 1 object (named \emph{Samp1}) of 393,338
  rows and 10 variables (the lengths and the distances are expressed in
  nucleotides):
\describe{
  \item{transcript}{Name of the transcript (ENST ID and version, dot
  separated)}
  \item{end5}{Position of the 5' end of the read with respect to the first
  nuclotide of the transcript}
  \item{psite}{Position of the P-site with respect to the first nuclotide of
  the transcript}
  \item{end3}{Position of the 3' end of the read with respect to the first
  nuclotide of the transcript}
  \item{length}{Length of the read}
  \item{start_pos}{Leftmost position of the CDS with respect to the first
  nuclotide of the transcript}
  \item{stop_pos}{Rightmost position of the CDS with respect to the first
  nuclotide of the transcript}
  \item{psite_from_start}{Position of the P-site with respect to the first
  nuclotide of the annotated CDS (if any)}
  \item{psite_from_stop}{Position of the P-site with respect to the last
  nuclotide of the annotated CDS (if any)}
  \item{psite_region}{Region of the transcript that includes the P-site
  (5utr, cds, 3utr)}
}}
\usage{
reads_psite_list
}
\description{
This dataset contains details on mapping reads after the identification of
the P-site and the update of \code{\link{reads_list}}.
}
\keyword{datasets}
