% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_plot.R
\name{rlength_distr}
\alias{rlength_distr}
\title{Read length distributions.}
\usage{
rlength_distr(data, sample, transcripts = NULL, cl = 100)
}
\arguments{
\item{data}{List of data tables from \code{\link{bamtolist}},
\code{\link{bedtolist}}, \code{\link{length_filter}} or
\code{\link{psite_info}}.}

\item{sample}{Character string specifying the name of the sample of interest.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.}

\item{cl}{Integer value in [1,100] specifying a confidence level for
restricting the plot to a sub-range of read lengths i.e. to the cl% of
read lengths associated to the highest signals. Default is 100.}
}
\value{
List containing a ggplot2 object ("plot") and the data table with the
  associated data ("dt").
}
\description{
This function generates read length distributions.
}
\examples{
data(reads_list)

## Generate the length distribution for all read lengths:
lendist_whole <- rlength_distr(reads_list, sample = "Samp1", cl = 100)
lendist_whole[["plot"]]

## Generate the length distribution for a sub-range of read lengths:
lendist_sub95 <- rlength_distr(reads_list, sample = "Samp1", cl = 95)
lendist_sub95[["plot"]]
}
