% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bam.R
\name{bamtobed}
\alias{bamtobed}
\title{Convert BAM files into BED files.}
\usage{
bamtobed(bamfolder, bedfolder = NULL)
}
\arguments{
\item{bamfolder}{A character string specifying the path to the BAM files.
The function looks for BAM format file recursively starting from the
specified folder.}

\item{bedfolder}{A character string specifying the (existing or not) location
of the output directory i.e. where the BED files shuold be saved. By
default this argument is NULL, which implies the folder is set as a
subfolder of \code{bamfolder}, called \emph{bed}.}
}
\description{
Converts a list of BAM files into a list of BED files containing, for each
read, the name of the transcript on which it alignes, the position of its
fist and last nucleotide, its length and the associated strand.
}
\examples{
path_bam <- location_of_BAM_files
path_bed <- location_of_output_directory
bamtobed(bamfolder = path_bam, bedfolder = path_bed)
}
