% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bam.R
\name{bamtolist}
\alias{bamtolist}
\title{Convert BAM files into a list of data frames or into a GRangesList object.}
\usage{
bamtolist(bamfolder, annotation, filter = "none", custom_range = NULL,
  periodicity_th = 50, list_name = NULL, granges = FALSE)
}
\arguments{
\item{bamfolder}{A character string indicating the path to the folder
containing the BAM files.}

\item{annotation}{A data frame from \code{\link{create_annotation}}. Please
make sure that the name of the reference sequences in the annotation data
frame coincides with those in the BAM files.}

\item{filter}{Either "none" (the default), "custom" or "periodicity". It
specifies how to handle the selection of the read length. "none": all read
lengths are included in the analysis; "custom": only read lengths specified
by the user are included (see \code{custom_range}); "periodicity": only
read lengths satisfying a periodicity threshold (see \code{periodicity_th})
are included in the analysis. This mode enables the removal of all the
reads without periodicity.}

\item{custom_range}{An integer or an integer vector specifying either the
read length or multiple read lengths to be kept, respectively. This
parameter is considered only when \code{filter} is set to "custom".}

\item{periodicity_th}{An integer in \emph{[10, 100]}. Only the read lengths
satisfying this threshold (i.e. with a higher percentage of read
extremities falling in the same frame) are kept. This parameter is
considered only when \code{filter} is set to "periodicity". Default is 50.}

\item{list_name}{A character string vector specifying the desired names for
the data frames of the output list. Its length must coincides with the
number of BAM files within \code{bamfolder}. Please pay attention to the
order in which they are provided: the first string is assigned to the first
file, the second string to the second one and so on. By default this
argument is NULL, implying that the data frames are named after the name of
the BAM file, leaving their path and extension out.}
}
\value{
A list of data frames or a GRangesList object.
}
\description{
Reads one or several BAM files, converts each file into a data frame and
combines them into a list. Alternatively, it returns a GRangesList i.e. a
list of GRanges objects. In both cases the data structure contains for each
read the name of the reference sequence (i.e. of the transcript) on which it
aligns, the leftmost and rightmost position of the read and its length. Two
additional columns are also attached, reporting the leftmost and rightmost
position of the CDS of the reference sequence with respect to its 1st
nuclotide. Please note: if a transcript is not associated to any annotated
CDS then its start and the stop codon are set to 0. Moreover, since the input
data refers to a transcriptome alingment, the reads mapping to the negative
strand should be present in a low percentage and they are removed. Multiple
options for treating the read lengths are available.
}
\examples{
path_bam <- location_of_BAM_files
annotation_df <- dataframe_with_transcript_annotation
bamtolist(bamfolder = path_bam, annotation = annotation_df)
}
