% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bam.R
\name{bedtolist}
\alias{bedtolist}
\title{Convert a list of BED files into a list of data frames.}
\usage{
bedtolist(bedfolder, annotation, list_name = NULL)
}
\arguments{
\item{bedfolder}{A character string indicating the folder where the BED files
are located.}

\item{annotation}{A data frame with a reference annotation of the transripts.
It must contain at least four columns named \emph{transcript},
\emph{l_utr5}, \emph{l_cds}, \emph{l_utr3} containing the name of the
transcripts (the same as in the reference transcriptome), the position of
the first nucleotide of the \emph{5' UTR}, the \emph{CDS} and the  \emph{3'
UTR}, respectively. No specific order is required.}

\item{list_name}{A character vector specifying the desired names to be 
assigned to the data frames of the output list. Its length must be equal to
the number of BED files within \code{bedfolder}. Pay attention to the order
in which their are provided: the first string will be assigned to the first
file, the second string to the second one and so on. By default this 
argument is NULL, which implies the data frames will be named as the BED 
files, leaving their path and extension out.}
}
\value{
A list of data frames.
}
\description{
Reads a list of BED files, converts each file into a data frame and combines
them into a list, so that each data frame in the list correspond to a sample.
Two additional columns are attached to the data frames, containing the
position of the start and the stop codons with respect to the beginning of
the transcript. Since the original BAM files come from an alingment on
transcripts, the reads associated to the negative strand should be present in
a low percentage, and they will be removed
}
\examples{
path_bed <- location_of_BED_files
annotation_df <- dataframe_with_transcript_annotation
bedtolist(bamfolder = path_bed, annotation = annotation_df)
}
