% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_bam.R
\name{bedtolist}
\alias{bedtolist}
\title{Convert a list of BED files into a list of data frames.}
\usage{
bedtolist(bedfolder, annotation, list_name = NULL)
}
\arguments{
\item{bedfolder}{A character string indicating the path to the folder where
the BED files are located.}

\item{annotation}{A data frame from \code{\link{create_annotation}}. Please
note that the transcript names in the annotation data frame should coincide
with those in the alignment file.}

\item{list_name}{A character vector specifying the desired names to be 
assigned to the data frames in the output list. Its length must be equal to
the number of BED files within \code{bedfolder}. Pay attention to the order
their are provided: the first string will be assigned to the first
file, the second string to the second one and so on. By default this 
argument is NULL, which implies the data frames will be named after the BED 
paths, leaving their path and extension out.}
}
\value{
A list of data frames.
}
\description{
Reads a list of BED files, converts each file into a data frame and combines 
them into a list, so that each data frame in the list correspond to a sample.
Two additional columns are attached to the data frames, containing the 
position of the start and the stop codon of the CDS with respect to the 1st
nuclotide of the transcript. Please note: if a transcript doesn't present any
annotated CDS then the positions of both the start and the stop codon will be
set to 0. Since the original BAM files come from a transcriptome alingment,
the reads aligned to the negative strand should be present in a low
percentage, and they will be removed.
}
\examples{
path_bed <- location_of_BED_files
annotation_df <- dataframe_with_transcript_annotation
bedtolist(bedfolder = path_bed, annotation = annotation_df)
}

