% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_coverage.R
\name{cds_coverage}
\alias{cds_coverage}
\title{Number of in-frame P-sites per coding sequence.}
\usage{
cds_coverage(data, annotation, start_nts = 0, stop_nts = 0)
}
\arguments{
\item{data}{List of data tables from \code{\link{psite_info}}.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{start_nts}{Positive integer specifying the number of nucleotides at the
beginning of the coding sequences to be excluded from the analisys. Default
is 0.}

\item{stop_nts}{Positive integer specifying the number of nucleotides at the
end of the coding sequences to be excluded from the analisys. Default is 0.}
}
\value{
A data table.
}
\description{
This function generates a data table that for each transcript contains at
least i) its name; ii) the length of its coding sequence; iii) the number of
in-frame P-sites falling in its annotated coding sequence (if any) for all
samples. A chosen number of nucleotides at the beginning and/or at the end of
the CDSs can be excluded for restricting the analysis to a portion of the
original coding sequence. In this case the output data table includes an
additional column reporting the length of the selected region. Please note:
transcripts without annotated CDS are automatically discarded.
}
\examples{
data(reads_psite_list)
data(mm81cdna)

## Compute the number of in-frame P-sites per whole coding sequence.
psite_cds <- cds_coverage(reads_psite_list, mm81cdna)

## Compute the number of in-frame P-sites per the coding sequence exluding
## the first 15 nucleotides and the last 10 nucleotides.
psite_cds <- cds_coverage(reads_psite_list, mm81cdna, start_nts = 15, stop_nts = 10)
}
