% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{codon_coverage}
\alias{codon_coverage}
\title{Compute the number of reads per codon.}
\usage{
codon_coverage(data, annotation, sample = NULL, psite = FALSE,
  min_overlap = 1, granges = FALSE)
}
\arguments{
\item{data}{A list of data tables from \code{\link{psite_info}}. Data tables
generated by \code{\link{bamtolist}} and \code{\link{bedtolist}} can be
used only if \code{psite} is FALSE (the default).}

\item{annotation}{A data table as generated by
\code{\link{create_annotation}}.}

\item{sample}{A character string vector specifying the name of the sample(s)
of interest. By default this argument is NULL, meaning that the coverage is
computed for all the samples in \code{data}.}

\item{psite}{A logical value whether or not to return the number of P-sites
per codon. Default is NULL, meaning that the number of read footprints per
codon is computed instead.}

\item{min_overlap}{A positive integer specyfing the minimum number of
overlapping positions (in nucleotides) between a reads and a codon to be
considered to be overlapping. When \code{psite} is TRUE this parameter must
be 1 (the default).}

\item{granges}{A logical value whether or not to return a GRanges object.
Default is FALSE, meaning that a data tables is returned instead.}
}
\value{
A data table or a GRanges object.
}
\description{
For the specified sample(s), this function computes the codon coverage
defined either as the number of read footprints per codon or as the number of
P-sites per codon.
}
\details{
The sequence of every transcript is divided in triplets starting
  from the annotated translation initiation site (if any) proceeding towards
  the UTRs extremities, and eventually discarding the exceeding 1 or 2
  nucleotides at the extremities of the transcript. Please note that the
  transcripts not associated to any annotated \emph{5' UTR}, \emph{CDS} and
  \emph{3'UTR} and transcripts with coding sequence length not divisible by 3
  are automatically discarded.
}
\examples{
data(reads_psite_list)
data(mm81cdna)

## Compute the coverage based on the number of ribosome footprint per codon,
## setting the minimum overlap between reads and triplets to 3 nts
## coverage_dt <- codon_coverage(reads_psite_list, mm81cdna, min_overlap = 3)

## Compute the coverage based on the number of P-sites per codon
## coverage_dt <- codon_coverage(reads_psite_list, mm81cdna, psite = TRUE)
}
