% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_usage_psite}
\alias{codon_usage_psite}
\title{Compute and plot empirical codon usage indexes based on the inferred P-sites.}
\usage{
codon_usage_psite(data, annotation, sample, fastapath = NULL,
  bsgenome_dp = NULL, txdb = NULL, transcripts = NULL,
  codon_usage = NULL, scatter_label = F, aminoacid = F)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}} that may or
may not include a column named \emph{psite_codon}. This column reports the
three nucleotides covered by the P-site and can be previously generated by
the \code{\link{psite_info}} function. If not already present, it is
automatically generated throughout the function starting from a FASTA file
or a BSgenome data package.}

\item{annotation}{A data frame as generated by
\code{\link{create_annotation}}.}

\item{sample}{A character string vector specifying the name of the sample of
interest.}

\item{fastapath}{A character string specifying the path to the FASTA file
containing the reference nucleotide sequences. Please make sure that the
sequences and their names derive from the same release of and are in
agreement with the annotation file used in the
\code{\link{create_annotation}} function. Either \code{fastapath} or
\code{bsgenome_dp} coulped with \code{txdb} must be specified if a column
named \emph{psite_codon} generated by \code{\link{psite_info}} is not
already present. Default is NULL.}

\item{bsgenome_dp}{A character string specifying the name of the BSgenome
data package to be loaded. If the specified data package is not already
present in your system, it is installed through the biocLite.R script.
Please check the data packages available in the Bioconductor repositories
for your version of R/Bioconductor using the
\code{\link[BSgenome]{available.genomes}} function from the BSgenome
package. Either \code{fastapath} or \code{bsgenome_dp} coulped with
\code{txdb} must be specified if a column named \emph{psite_codon}
generated by \code{\link{psite_info}} is not already present. Default is
NULL.}

\item{txdb}{A TxDb object. This parameter is considered only if
\code{bsgenome_dp} is specified. Default is NULL.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the analysis. By default this argument is
NULL, meaning that all the transcripts in \code{data} are used. Please note
that the transcripts not associated to any annotated \emph{5' UTR},
\emph{CDS} and \emph{3'UTR} and transcripts with coding sequence length not
divisible by 3 are automatically discarded.}

\item{codon_usage}{A data frame containing codon usage values provided by the
user. These values are compared with the empirical codon usage indexes of
the sample of interest. The data frame must contain at least the 64 codons
and the corresponding values arranged in two columns named \emph{codon} and
\emph{usage_index}, respectively. Note that a similar data frame is also
returned by \code{\link{codon_usage_psite}} itself. Default is NULL.}

\item{scatter_label}{A logical value whether or not to label the dots of the
scatter plot generated by specifying \code{codon_usage}. Each dot can be
labeled either after the three nucleotides of the codon or after the
corresponding amino acid (see \code{aminoacid}). This parameter is
considered only if \code{codon_usage} is specified. Default is FALSE.}

\item{aminoacid}{A logical value whether or not to label the dots of the
scatter plot generated by specifying \code{codon_usage} using the amino
acids corresponding to the triplets. Default is FALSE, meaning that the
three nucleotides of the codon are used instead. This parameter is
considered only if \code{codon_usage} is specified and \code{scatter_label}
is TRUE. Default is FALSE.}
}
\value{
A list containing a ggplot2 object (named "plot"), and a data frame
  ("df") with the associated data. An additional ggplot2 object
  ("plot_comparison") is returned if \code{codon_usage} is specified.
}
\description{
For a specified sample this function computes an empirical codon usage index
based on the frequency of in-frame P-sites along the coding sequence. It
computes the codon usage index for all the 64 triplets, normalizes them for
the frequency of each codon within the CDS and returns a bar plot with the
resulting values. This function also allows to compare the computed codon
usage indexes with a set of 64 values provided by the user.
}
