% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_usage_psite}
\alias{codon_usage_psite}
\title{Compute and plot the usage of the codons.}
\usage{
codon_usage_psite(data, annotation, sample, fastapath, bsgenome_dp = NULL,
  txdb = NULL, transcripts = NULL, codon_usage = NULL,
  scatter_label = F, aminoacid = F)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}} that may or 
may not include a column \emph{psite_codon}, reporting the nucleotide
sequence of the P-sites. This column can be generated by running 
\code{\link{psite_info}} specifying the \code{fastapath} option. If the 
data frame of interest does not include a \emph{psite_codon} column, the 
information about the nucleotide sequnce of the P-sites will be 
automatically generated throughout the function starting from the FASTA 
specified in \code{fastapath}.}

\item{annotation}{A data frame from \code{\link{create_annotation}}.}

\item{sample}{A character string vector specifying the name of the sample
of interest.}

\item{fastapath}{A character string specifying the path to the FASTA file
containing the nucleotide sequence of the transcripts. For each mRNA the
the record description line must contain the transcript name as in the
reference transcriptome and the sequence must derive from the same relase
of the genome. Either \code{fastapath} or \code{bsgenome_dp} coulped with
\code{txdb} must be specified if the data frame of interest does not
already include a \emph{psite_codon} column (see \code{\link{psite_info}}).
This arguments are used to extract for each read the three nucletotides
covered by the identified P-sites.}

\item{bsgenome_dp}{A character string specifying the name of the BSgenome
data package to be loaded. If the specified data package is not already
present in your system, it will be installed through the biocLite.R script.
Please check the data packages available in the Bioconductor repositories
for your version of R/Bioconductor using the
\code{\link[BSgenome]{available.genomes}} function from the BSgenome package.
Either \code{fastapath} or \code{bsgenome_dp} coulped with
\code{txdb} must be specified if the data frame of interest does not
already include a \emph{psite_codon} column (see \code{\link{psite_info}}).
This arguments are used to extract for each read the three nucletotides
covered by the identified P-sites.}

\item{txdb}{A TxDb object storing transcript annotations.  It is considered
only if \code{bsgenome_dp} is specified.}

\item{transcripts}{A character string vector specifying the name of the 
transcripts to be considered in the analysis. By default this argument is 
NULL, which implies all the transcripts in \code{data} will be used. Either
transcripts with no annotated \emph{5' UTR}, \emph{CDS} and
\emph{3'UTR} or mRNAs not labeled as protein coding will be removed.}

\item{codon_usage}{A data frame containing codon usage values provided by the
user. These values will be compared with the codon usage of the sample of
interest. The data frame must contain at least the 64 codons and the
corresponding codon usage values, arranged in two columns named
\emph{codon} and \emph{usage_index}, respectively. A data frame including
these information is returned by the \code{\link{codon_usage_psite}}
function. Default is NULL, meaning that no comparisons will be performed.}

\item{scatter_label}{A logical value whether or not to label the point of the
scatter plot based on \code{codon_usage}. It is possible to associate to 
each point the triplet of nucleotides or the corresponding amino acid (see
\code{aminoacid}). Default is FALSE.}

\item{aminoacid}{A logical value whether or not to label the point of the 
scatter plot based on \code{codon_usage} using the amino acids 
corresponding to the triplets. Default is FALSE, meaning that the codon
sequences will be used instead. It is considered only if 
\code{scatter_label} is TRUE.}
}
\value{
A list containing a ggplot2 object (named "plot"), and a data frame
  ("df") with the associated data. An additional ggplot2 object
  ("plot_comparison") is returned if the parameter \code{codon_usage} is
  specified.
}
\description{
For a specified sample computes the empirical codon usage i.e. the frequency 
of in-frame P-sites along the coding sequence codon by codon, normalized for 
the frequency in sequences of each codon. This function also allows to
compare the empirical codon usage of the specified sample with a set of
values provided by the user.
}
