% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_usage_psite}
\alias{codon_usage_psite}
\title{Compute and plot empirical codon usage indexes based on the inferred P-sites.}
\usage{
codon_usage_psite(data, annotation, sample, fastapath = NULL,
  fasta_genome = TRUE, bsgenome = NULL, gtfpath = NULL, txdb = NULL,
  dataSource = NA, organism = NA, transcripts = NULL,
  codon_values = NULL, scatter_label = FALSE, aminoacid = FALSE)
}
\arguments{
\item{data}{A list of data tables from \code{\link{psite_info}} that may or
may not include a column named \emph{psite_codon}. This column reports the
three nucleotides covered by the P-site and can be previously generated by
the \code{\link{psite_info}} function. If not already present, it is
automatically generated throughout the function starting from a FASTA file
or a BSgenome data package.}

\item{annotation}{A data table as generated by
\code{\link{create_annotation}}.}

\item{sample}{A character string vector specifying the name of the sample of
interest.}

\item{fastapath}{An optional character string specifying the path to the
FASTA file used in the alignment step, including its name and extension.
This file can contain reference nucleotide sequences either of a genome
assembly or of all the transcripts (see \code{fasta_genome}). Please make
sure the sequences derive from the same release of the annotation file used
in the \code{\link{create_annotation}} function. Note: either
\code{fastapath} or \code{bsgenome} is required to normalize the data, even
if a column \emph{psite_codon} has been previously generated by
\code{\link{psite_info}}. Default is NULL.}

\item{fasta_genome}{A locigal value whether or not the FASTA file specified
by \code{fastapath} contains nucleotide sequences of a genome assembly.
FALSE means that the nucleotide sequences of all the transcripts are
provided instead. When this parameter is TRUE (the default), an annotation
object is required (see \code{gtfpath} and \code{txdb}).}

\item{bsgenome}{An optional character string specifying the name of the
BSgenome data package containing the genome sequences to be loaded. If it
is not already present in your system, it will be installed through the
biocLite.R script (check the list of data packages available in the
Bioconductor repositories for your version of R/Bioconductor by the
\code{\link[BSgenome]{available.genomes}} function of the BSgenome
package). This parameter also requires an annotation object (see
\code{gtfpath} and \code{txdb}). Please make sure the sequences included in
the specified BSgenome data pakage are in agreement with the sequences used
in the alignment step. Note: either \code{fastapath} or \code{bsgenome} is
required to normalize the data, even if a column \emph{psite_codon} has
been previously generated by \code{\link{psite_info}}. Default is NULL.}

\item{gtfpath}{A character string specifying the path to te GTF file,
including its name and extension. Please make sure the GTF derives from the
same release of what is specified by \code{fastapath} or by
\code{bsgenome}. Note that either \code{gtfpath} or \code{txdb} must be
specified when the nucleotide sequences of a genome assembly are provided
(see \code{fastapath} or \code{bsgenome}). Default is NULL.}

\item{txdb}{A character string specifying the name of the annotation package
for TxDb object(s) to be loaded. If it is not already present in your
system, it will be installed through the biocLite.R script (check the list
of TxDb annotation packages available in the Bioconductor repositories at
http://bioconductor.org/packages/release/BiocViews.html#___TxDb )). Please
make sure the annotation package derives from the same release of what is
specified by \code{fastapath} or by \code{bsgenome}. Note that either
\code{gtfpath} or \code{txdb} must be specified when the nucleotide
sequences of a genome assembly are provided (see \code{fastapath} or
\code{bsgenome}). Default is NULL.}

\item{dataSource}{An optional character string describing the origin of the
GTF data file. For more information about this parameter please refer to
the description of \emph{dataSource} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}

\item{organism}{A optional character string reporting the genus and species
of the organism when \code{gtfpath} is specified.  For more information
about this parameter please refer to the description of \emph{dataSource}
of the \code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in
the \code{GenomicFeatures} package.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the analysis. By default this argument is
NULL, meaning that all the transcripts in \code{data} are used. Please note
that the transcripts not associated to any annotated \emph{5' UTR},
\emph{CDS} and \emph{3'UTR} and transcripts with coding sequence length not
divisible by 3 are automatically discarded.}

\item{codon_values}{A data table containing codon-specific values provided by
the user. These values are compared with the empirical codon usage indexes
of the sample of interest. The data table must contain at least the 64
codons and the corresponding values arranged in two columns named
\emph{codon} and \emph{value}, respectively. Note that a similar data
table is also returned by \code{\link{codon_usage_psite}} itself. Default
is NULL.}

\item{scatter_label}{A logical value whether or not to label the dots of the
scatter plot generated by specifying \code{codon_values}. Each dot can be
labeled either after the three nucleotides of the codon or after the
corresponding amino acid (see \code{aminoacid}). This parameter is
considered only if \code{codon_values} is specified. Default is FALSE.}

\item{aminoacid}{A logical value whether or not to label the dots of the
scatter plot generated by specifying \code{codon_values} using the amino
acids corresponding to the triplets. Default is FALSE, meaning that the
three nucleotides of the codon are used instead. This parameter is
considered only if \code{codon_values} is specified and \code{scatter_label}
is TRUE. Default is FALSE.}
}
\value{
A list containing a ggplot2 object (named "plot"), and a data table
  ("dt") with the associated data. An additional ggplot2 object
  ("plot_comparison") is returned if \code{codon_values} is specified.
}
\description{
For a specified sample this function computes an empirical codon usage index
based on the frequency of in-frame P-sites along the coding sequence. It
computes the codon usage index for all the 64 triplets, normalizes them for
the frequency of each codon within the CDS and returns a bar plot with the
resulting values. This function also allows to compare the computed codon
usage indexes with a set of 64 values provided by the user.
}
