% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_usage_psite}
\alias{codon_usage_psite}
\title{Empirical codon usage indexes.}
\usage{
codon_usage_psite(data, annotation, sample, site = "psite",
  fastapath = NULL, fasta_genome = TRUE, bsgenome = NULL,
  gtfpath = NULL, txdb = NULL, dataSource = NA, organism = NA,
  transcripts = NULL, codon_values = NULL, scatter_label = FALSE,
  aminoacid = FALSE)
}
\arguments{
\item{data}{List of data tables from \code{\link{psite_info}}. Each data
table may or may not include one or more columns among \emph{p_site_codon},
\emph{a_site_codon} and \emph{e_site_codon} reporting the three nucleotides
covered by the P-site, A-site and E-site, respectively. These columns can
be previously generated by the \code{\link{psite_info}} function. Otherwise
the column of interest can be specified by \code{site} and is automatically
generated starting from a FASTA file or a BSgenome data package.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Character string specifying the name of the sample of interest.}

\item{site}{Either "psite, "asite", "esite". It specifies if the empirical
codon usage indexes should be based on ribosome P-sites ("psite"), A-sites
("asite") or E-sites ("esite"). Default is "psite".}

\item{fastapath}{Character string specifying the FASTA file used in the
alignment step, including its path, name and extension. This file can
contain reference nucleotide sequences either of a genome assembly or of
all the transcripts (see \code{Details} and \code{fasta_genome}). Please
make sure the sequences derive from the same release of the annotation file
used in the \code{\link{create_annotation}} function. Note: either
\code{fastapath} or \code{bsgenome} is required to compute the codon
frequencies within the CDS used as normalization factors, even when
\code{data} already includes one or more columns among \emph{p_site_codon},
\emph{a_site_codon} and \emph{e_site_codon}. Default is NULL.}

\item{fasta_genome}{Logical value whether the FASTA file specified by
\code{fastapath} contains nucleotide sequences of a genome assembly. If
TRUE (the default), an annotation object is required (see \code{gtfpath}
and \code{txdb}). FALSE implies the nucleotide sequences of all the
transcripts is provided instead.}

\item{bsgenome}{Character string specifying the BSgenome data package with
the genome sequences to be loaded. If not already present in the system, it
is automatically installed through the biocLite.R script (check the list of
available BSgenome data packages by the
\code{\link[BSgenome]{available.genomes}} function of the BSgenome
package). This parameter must be coupled with an annotation object (see
\code{gtfpath} and \code{txdb}). Please make sure the sequences included in
the specified BSgenome data pakage are in agreement with the sequences used
in the alignment step. Note: either \code{fastapath} or \code{bsgenome} is
required to compute the codon frequencies within the CDS used as
normalization factors, even when \code{data} already includes one or more
columns among \emph{p_site_codon}, \emph{a_site_codon} and
\emph{e_site_codon}. Default is NULL.}

\item{gtfpath}{Character string specifying the location of a GTF file,
including its path, name and extension. Please make sure the GTF file and
the sequences specified by \code{fastapath} or \code{bsgenome} derive from
the same release. Note that either \code{gtfpath} or \code{txdb} is
required if and only if nucleotide sequences of a genome assembly are
provided (see \code{fastapath} or \code{bsgenome}). Default is NULL.}

\item{txdb}{Character string specifying the TxDb annotation package to be
loaded. If not already present in the system, it is automatically installed
through the biocLite.R script (check
\href{http://bioconductor.org/packages/release/BiocViews.html#___TxDb}{here}
 the list of available TxDb annotation packages). Please make sure the TxDb
annotation package and the sequences specified by \code{fastapath} or
\code{bsgenome} derive from the same release. Note that either
\code{gtfpath} or \code{txdb} is required if and only if nucleotide
sequences of a genome assembly are provided (see \code{fastapath} or
\code{bsgenome}). Default is NULL.}

\item{dataSource}{Optional character string describing the origin of the GTF
data file. This parameter is considered only if \code{gtfpath} is
specified. For more information about this parameter please refer to the
description of \emph{dataSource} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}

\item{organism}{Optional character string reporting the genus and species of
the organism of the GTF data file. This parameter is considered only if
\code{gtfpath} is specified. For more information about this parameter
please refer to the description of \emph{organism} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.
Please note: transcripts without annotated CDS and transcripts whose coding
sequence length is not divisible by 3 are automatically discarded.}

\item{codon_values}{Data table containing 64 codon-specific values. If
specified, the provided values are compared with the empirical codon usage
indexes computed for the sample of interest. The data table must contain
the DNA or RNA nucleotide sequence of the 64 codons and the corresponding
values arranged in two columns named \emph{codon} and \emph{value},
respectively. Please note: a data table of the same format is returned by
\code{\link{codon_usage_psite}} itself. Default is NULL.}

\item{scatter_label}{Logical value whether to label the dots of the scatter
plot generated by specifying \code{codon_values}. Each dot is labeled using
either the nucleotide sequence of the codon or the corresponding amino acid
symbol (see \code{aminoacid}). This parameter is considered only if
\code{codon_values} is specified. Default is FALSE.}

\item{aminoacid}{Logical value whether to use the amino acid symbols to label
the dots of the scatter plot generated by specifying \code{codon_values}.
Default is FALSE i.e. the nucleotide sequences of the codons are used
instead. This parameter is considered only if \code{codon_values} is
specified and \code{scatter_label} is TRUE. Default is FALSE.}
}
\value{
A list containing a ggplot2 object ("plot") and the data table with
  the associated data ("dt"). If \code{codon_values} is specified, an
  additional ggplot2 object ("plot_comparison") is returned.
}
\description{
This function computes empirical codon usage indexes based on either ribosome
P-sites, A-site or E-site frequencies associated to in-frame P-sites within
the coding sequence. It computes 64 codon usage indexes (one per triplet)
normalized for the frequency of the corresponding codons within the CDS and
generates a bar plot of the resulting values. Optionally, this function
compares the computed codon usage indexes with a set of 64 values provided by
the user. In this case the function returns a scatter plot, reporting the
result of a linear regression between the two variables (i.e. the two sets
of values) and the corresponding Pearson correlation coefficient.
}
\details{
\strong{riboWaltz} only works for read alignments based on
  transcript coordinates. This choice is due to the main purpose of RiboSeq
  assays to study translational events through the isolation and sequencing
  of ribosome protected fragments. Most reads from RiboSeq are supposed to
  map on mRNAs and not on introns and intergenic regions. Nevertheless, BAM
  based on transcript coordinates can be generated in two ways: i) aligning
  directly against transcript sequences; ii) aligning against standard
  chromosome sequences, requiring the outputs to be translated in transcript
  coordinates. The first option can be easily handled by many aligners (e.g.
  Bowtie), given a reference FASTA file where each sequence represents a
  transcript, from the beginning of the 5' UTR to the end of the 3' UTR. The
  second procedure is based on reference FASTA files where each sequence
  represents a chromosome, usually coupled with comprehensive gene annotation
  files (GTF or GFF). The STAR aligner, with its option --quantMode
  TranscriptomeSAM (see Chapter 6 of its
  \href{http://labshare.cshl.edu/shares/gingeraslab/www-data/dobin/STAR/STAR.posix/doc/STARmanual.pdf}{manual}),
   is an example of tool providing such a feature.
}
