% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{create_annotation}
\alias{create_annotation}
\title{Create an annotation data frame.}
\usage{
create_annotation(gtfpath = NULL, txdb = NULL, dataSource = NA,
  organism = NA)
}
\arguments{
\item{gtfpath}{A character string specifying the path to te GTF file,
including its name and extension. Either \code{gtfpath} or \code{txdb} must
be specified.}

\item{txdb}{A TxDb object. Either \code{gtfpath} or \code{txdb} must be
specified.}

\item{dataSource}{A character string describing the origin of the data file.
For more information about this parameter please refer to the description
of \emph{dataSource} of the \code{\link[GenomicFeatures]{makeTxDbFromGFF}}
function included in the \code{GenomicFeatures} package.}

\item{organism}{A character string specifying the Genus and species of this
organism. For more information about this parameter please refer to the
description of \emph{dataSource} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}
}
\value{
A data frame.
}
\description{
Starting from a GTF file or a TxDb object this function generates a dada
frame containing a basic annotation of the transcripts. The data frame
includes a column named \emph{transcript} reporting the name of the reference
sequences and four columns named \emph{l_tr}, \emph{l_utr5}, \emph{l_cds} and
\emph{l_utr3} reporting the length of the transcripts and of their annotated
\emph{5' UTR}, \emph{CDS} and \emph{3' UTR}, respectively.
}
\examples{
gtf_file <- location_of_GTF_file
path_bed <- location_of_output_directory
bamtobed(gtfpath = gtf_file, dataSource = "gencode6", organism = "Mus musculus")
}
