% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{create_annotation}
\alias{create_annotation}
\title{Create an annotation data frame.}
\usage{
create_annotation(gtfpath = NULL, txdb = NULL, dataSource = NA,
  organism = NA)
}
\arguments{
\item{gtfpath}{A character string specifying the complete path to te GTF
file, including its name and extension. Either \code{gtfpath} or
\code{txdb} must be specified.}

\item{txdb}{A TxDb object storing transcript annotations. Either
\code{gtfpath} or \code{txdb} must be specified.}

\item{dataSource}{A character string describing the origin of the data file.
Please refer to the description of \emph{dataSource} of the
\code{makeTxDbFromGFF} function included in the \code{GenomicFeatures}
package.}

\item{organism}{A character string specifying the Genus and species of this
organism. Please refer to the description of  \emph{organism} of the
\code{makeTxDbFromGFF} function included in the \code{GenomicFeatures}
package.}
}
\value{
A data frame.
}
\description{
Creates a data frame with a basic reference annotation of the transripts 
starting from a GTF file. The data frame will contain a column named 
\emph{transcript} reporting the name of the transcripts (as reported in the 
GTF, usually the ENST ID and version, dot separated) and four columns named 
\emph{l_tr}, \emph{l_utr5}, \emph{l_cds} and \emph{l_utr3} reporting the 
length of the whole transcript and of the annotated \emph{5' UTR}, the 
\emph{CDS} and the \emph{3' UTR}, respectively. Please note that the 
transcript names should coincide with those in the alignment file.
}
\examples{
gtf_file <- location_of_GTF_file
path_bed <- location_of_output_directory
bamtobed(gtfpath = gtf_file, dataSource = "gencode6", organism = "Mus musculus")
}
