% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{create_annotation}
\alias{create_annotation}
\title{Create an annotation data frame.}
\usage{
create_annotation(gtfpath = NULL, txdb = NULL, dataSource = NA,
  organism = NA)
}
\arguments{
\item{gtfpath}{A character string specifying the path to te GTF file,
including its name and extension. Please make sure the GTF derives from the
same release of the sequences used in the alignment step. Note that either
\code{gtfpath} or \code{txdb} must be specified.}

\item{txdb}{A character string specifying the name of the annotation package
for TxDb object(s) to be loaded. If it is not already present in your
system, it will be installed through the biocLite.R script (check the list
of TxDb annotation packages available in the Bioconductor repositories at
http://bioconductor.org/packages/release/BiocViews.html#___TxDb )). Please
make sure the annotation package derives from the same release of the
sequences used in the alignment step. Note that either \code{gtfpath} or
\code{txdb} must be specified.}

\item{dataSource}{An optional character string describing the origin of the
GTF data file. For more information about this parameter please refer to
the description of \emph{dataSource} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}

\item{organism}{A optional character string reporting the genus and species
of the organism when \code{gtfpath} is specified.  For more information
about this parameter please refer to the description of \emph{dataSource}
of the \code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in
the \code{GenomicFeatures} package.}
}
\value{
A data frame.
}
\description{
Starting from a GTF file or a TxDb object this function generates a dada
frame containing a basic annotation of the transcripts. The data frame
includes a column named \emph{transcript} reporting the name of the reference
sequences and four columns named \emph{l_tr}, \emph{l_utr5}, \emph{l_cds} and
\emph{l_utr3} reporting the length of the transcripts and of their annotated
\emph{5' UTR}, \emph{CDS} and \emph{3' UTR}, respectively.
}
\examples{
gtf_file <- location_of_GTF_file
path_bed <- location_of_output_directory
bamtobed(gtfpath = gtf_file, dataSource = "gencode6", organism = "Mus musculus")
}
