% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_psite}
\alias{frame_psite}
\title{Percentage of P-sites per reading frame.}
\usage{
frame_psite(
  data,
  sample = NULL,
  transcripts = NULL,
  region = "all",
  length_range = "all",
  plot_title = NULL
)
}
\arguments{
\item{data}{List of data tables from \code{\link{psite_info}}.}

\item{sample}{Character string vector specifying the name of the sample(s) of
interest. Default is NULL i.e. all samples in \code{data} are processed.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.}

\item{region}{Character string specifying the region(s) of the transcripts to
be analysed. It can be either "5utr", "cds", "3utr" for 5' UTRs, CDSs and
3' UTRs, respectively. Default is "all" i.e. all regions are considered.
According to this parameter the bar plots are differently arranged to
optimise the organization and the visualization of the data.}

\item{length_range}{Integer or an integer vector specyfying the read
length(s) to be included in the analysis. Default is "all" i.e. all read
lengths are used.}

\item{plot_title}{Character string specifying the title of the plot. If
"auto", the title of the plot reports the region specified by \code{region}
(if any) and the considered read length(s). Default is NULL i.e. no title
is plotted.}
}
\value{
A list containing a ggplot2 object ("plot") and the data table with
  the associated data ("dt").
}
\description{
This function computes the percentage of P-sites falling in the three
possible translation reading frames and generates a bar plot of the resulting
values. It only handles annotated 5' UTRs, coding sequences and 3' UTRs,
separately.
}
\examples{
## data(reads_list)
##
## ## compute and add p-site datails
## psite_offset <- psite(reads_list, flanking = 6, extremity = "auto")
## reads_psite_list <- psite_info(reads_list, psite_offset)
##
## ## Generate the bar plot for all read lengths:
## frame_whole <- frame_psite(reads_psite_list, sample = "Samp1")
## 
## ## Generate the bar plot restricting the analysis to coding sequences and
## ## reads of 28 nucleotides:
## frame_sub <- frame_psite(reads_psite_list, sample = "Samp1",
##                          region = "cds", length_range = 28)
}
