% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_psite}
\alias{frame_psite}
\title{Compute the number of P-sites per frame.}
\usage{
frame_psite(data, sample = NULL, region = "all", length_range = "all")
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{sample}{A list of character string vectors specifying the name of the
samples of interest. Default is NULL, meaning that all the samples will be
taken into account.}

\item{region}{Either "all" or a character string among "5utr", "cds", "3utr"
specifying which of the three regions of the transcript (5' UTR, CDS or 3'
UTR, respectively) should be considered. Default is "all", meaning that the
percentage is computed and plotted for all the three regions.}

\item{length_range}{Either "all", an integer or an integer vector. In the
first case all the reads will be used to generate barplot. Otherwise, only
the reads matching the specified length(s) will be employed.}
}
\value{
A list containing a ggplot2 object and a data frame with the
  associated data.
}
\description{
Generates a barplot reporting the percentage of identified P-sites on the 
three reading frames of the transcript. This value is computed for the
annotated 5' UTRs, coding sequences and 3' UTRs, separately.
}
\examples{
data(reads_psite_list)

## Generate the barplot employing the whole dataset
frame_whole <- frame_psite(reads_psite_list, sample = "Samp1")
frame_whole[["plot"]]

## Generate the barplot for the coding sequence employing reads of 28
nucleotides
frame_sub <- frame_psite(reads_psite_list, sample = "Samp1", region = "cds",
length_range=28)
frame_sub[["plot"]]
}
