% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_psite_length}
\alias{frame_psite_length}
\title{Compute the number of P-sites per frame stratified by read length.}
\usage{
frame_psite_length(data, sample = NULL, region = "all", cl = 100,
  length_range = NULL)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{sample}{A character string vector specifying the name of the sample(s)
of interest. By default this argument is NULL, meaning that all the samples
in \code{data} are included in the analysis.}

\item{region}{Either "all" or a character string among "5utr", "cds", "3utr"
specifying the regions of the transcript (5' UTR, CDS or 3' UTR,
respectively) that must be included in the analysis. Default is "all",
meaning that the all the regions are considered.}

\item{cl}{An integer value in \emph{[1,100]} specifying the confidence level
for restricting the analysis to a sub-range of read lengths. By default it is
set to 100. This parameter has no effect if \code{length_range} is
specified.}

\item{length_range}{Either "all", an integer or an integer vector. Default is
"all", meaning that all the read lengths will be included in the analysis.
Otherwise, only the read lengths matching the specified value(s) are kept.
If specified, this parameter prevails over \code{cl}.}
}
\value{
A list containing a ggplot2 object and a data frame with the
  associated data.
}
\description{
Similar to \code{\link{frame_psite}} but the results are stratified by the
length of the reads.
}
\examples{
data(reads_psite_list)

## Generate the heatmap for all the read lengths
frame_len_whole <- frame_psite_length(reads_psite_list, sample = "Samp1")
frame_len_whole[["plot"]]

## Generate the heatmap for a sub-range of read lengths (the middle 90\%) and
restricting the analysis to the coding sequence
frame_len_sub <- frame_psite_length(reads_psite_list, sample = "Samp1",
region = "cds", cl = 90)
frame_len_sub[["plot"]]
}
