% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_psite_length}
\alias{frame_psite_length}
\title{Compute the number of P-sites per frame for each read length.}
\usage{
frame_psite_length(data, sample = NULL, region = "all", cl = 100,
  length_range = NULL)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{region}{Either "all" or a character string among "5utr", "cds", "3utr"
specifying which of the three regions of the transcript (5' UTR, CDS or 3'
UTR, respectively) should be considered. Default is "all", meaning that the
percentage is computed and plotted for all the three regions.}

\item{cl}{An integer with value in \emph{[1,100]} specifying the read length
confidence level for restricting the distribution to a chosen range of
lengths. By default it is set to 100, i.e. the whole set of reads lengths
will be considered. This parameter has no effect if \code{length_range} is
specified.}

\item{length_range}{Either "all", an integer or an integer vector. In the
first case all the reads will be used to generate barplot. Otherwise, only
the reads matching the specified length(s) will be employed. If specified,
this parameter prevails over \code{cl}}

\item{A}{list of character string vectors specifying the name of the samples
of interest. Default is NULL, meaning that all the samples will be taken
into account.}
}
\value{
A list containing a ggplot2 plot object and a data frame with the
  associated data.
}
\description{
Similar to \code{\link{frame_psite}} but in this case the
results are stratified by the length of the reads.
}
\examples{
data(reads_psite_list)

## Generate the heatmap employing the whole dataset
frame_len_whole <- frame_psite_length(reads_psite_list, sample = "Samp1")
frame_len_whole[["plot"]]

## Generate the heatmap for the coding sequence and visualize it for the
middle 98\% of the range of read length
frame_len_sub <- frame_psite_length(reads_psite_list, sample = "Samp1",
region = "cds", cl = 98)
frame_len_sub[["plot"]]
}
