% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_filter.R
\name{length_filter}
\alias{length_filter}
\title{Read length filtering.}
\usage{
length_filter(
  data,
  length_filter_mode,
  length_filter_vector = NULL,
  periodicity_threshold = 50,
  granges = FALSE
)
}
\arguments{
\item{data}{List of data tables from \code{\link{bamtolist}},
\code{\link{bedtolist}} or \code{\link{psite_info}}.}

\item{length_filter_mode}{Either "custom" or "periodicity". It specifies how
read length selection should be performed. "custom": only read lengths
specified by the user are kept (see \code{length_filter_vector});
"periodicity": only read lengths satisfying a periodicity threshold (see
\code{periodicity_threshold}) are kept. The latter mode enables the removal
of all reads with low or no periodicity.}

\item{length_filter_vector}{Integer or integer vector specifying one read
length or a range of read lengths to keep, respectively. This parameter is
considered only if \code{length_filter_mode} is set to "custom".}

\item{periodicity_threshold}{Integer in [10, 100]. Only read lengths
satisfying this threshold (i.e. a higher percentage of read extremities
falls in one of the three reading frames along the CDS) are kept. This
parameter is considered only if \code{length_filter_mode} is to
"periodicity". Default is 50.}

\item{granges}{Logical value whether to return a GRangesList object. Default
is FALSE i.e. a list of data tables is returned instead (the required input
for \code{\link{psite}}, \code{\link{psite_info}}, \code{\link{rends_heat}}
and \code{\link{rlength_distr}}).}
}
\value{
A list of data tables or a GRangesList object.
}
\description{
Read length filtering.
}
\examples{
data(reads_list)

## Keep reads of length between 27 and 30 nucleotides (included):
filtered_list <- length_filter(reads_list, length_filter_mode = "custom",
length_filter_vector = 27:30)

## Keep reads of lengths satisfying a periodicity threshold (70\%):
filtered_list <- length_filter(reads_list, length_filter_mode = "periodicity",
periodicity_threshold = 70)
}
