% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaheatmap_psite}
\alias{metaheatmap_psite}
\title{Plot ribosome occupancy metaheatmap at single-nucleotide resolution.}
\usage{
metaheatmap_psite(data, annotation, sample, scale_factors = NULL,
  length_range = "all", transcripts = NULL, utr5l = 25, cdsl = 50,
  utr3l = 25, log = F, colour = "black", plot_title = NULL)
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{annotation}{A data frame with a reference annotation of the transripts.
It must contain at least four columns named \emph{transcript},
\emph{l_utr5}, \emph{l_cds}, \emph{l_utr3} containing the name of the
transcripts (the same as in the reference transcriptome), the position of
the first nucleotide of the \emph{5' UTR}, the \emph{CDS} and the  \emph{3'
UTR}, respectively. No specific order is required.}

\item{sample}{A list of character string vectors specifying the name of the
samples of interest. The samples contained in each elements of the list
will be merged toghether, using the scale factors specified by
\code{scale_factors} (if any). The name of the list elements will be used
as labels for the raws of the heatmaps.}

\item{scale_factors}{A numeric vector of scale factors to be used for merging
the replicates of the specified sample (if any). The vector must contain at
least a set of values named after the strings listed in \code{sample}. No
specific order is required.}

\item{length_range}{Either "all", an integer or an integer vector. In the
first case all the reads will be used to generate the metaheatmap.
Otherwise, only the reads matching the specified length(s) will be
employed.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the metaprofile. By default this argument is
NULL, which implies all the transcripts in \code{data} will be used. Note
that if either the 5' UTR, the coding sequence or the 3' UTR of a
transcript is shorther than what is specified by \code{utr5l},
\eqn{2*}\code{cdsl} and \code{utr3l} respectively, the transcript will not
be cosidered.}

\item{utr5l}{A positive integer specifying the length (in nucleotides) of the
5' UTR portion that will flank the start codon in the plot. The default
value is 25.}

\item{cdsl}{A positive integer specifying the length (in nucleotides) of the
coding sequence portion that will flank both the start and the stop codon
in the plot. The default value is 50.}

\item{utr3l}{A positive integer specifying the length (in nucleotides) of the
3' UTR portion that will flank the stop codon in the plot. The default
value is 25.}

\item{log}{A logical value whether or not to use a logarithmic scale colour
(it is suggested for data with a strong difference between the lowest and
the highest signal. Default is FALSE.}

\item{colour}{A character string specifying the colour to be used for the
plot.}

\item{plot_title}{Either NULL (the default), "auto" or any character string.
When \emph{auto}, the title of the plot reports the number of transcript
and the read lengths employed for generating the heatmap.}
}
\value{
A list containing a ggplot2 plot object, a data frame with the
  associated data and the transcripts employed for generating the plot.
}
\description{
Plots a heatmap-like metaprofile based on the identified P-site of the reads.
It is similar to \code{\link{metaprofile_psite}}, but here the intensity of
the signal along the chosen region is represented by a continuous color scale
rather than by the height of a line chart. This graphical output is the
optimal choice in order to compare multiple samples or look at the behaviour
of the data if different populations of reads are considered.
}
\examples{
data(reads_psite_list)

## Generate the metaheatmap employing the whole dataset
metaheat_whole <- metaheatmap_psite(reads_psite_list, mm81cdna, sample = list("Whole"=c("Samp1")))
metaheat_whole[["plot"]]

## Generate the metaheatmap employing reads of 27, 28 and 29 nucleotides and
a specified set of transcripts (for example those with at least one P-site on
the first nucleotide of the coding sequence)
sample_name <- "Samp1"
sub_reads_psite_list <- subset(reads_psite_list[[sample_name]], psite_from_start == 0)
transcript_names <- as.character(sub_reads_psite_list$transcript)
metaheat_sub <- metaheatmap_psite(reads_psite_list, mm81cdna, sample = list("sub"=sample_name),
length_range = 27:29, transcripts = transcript_names, plot_title = "auto")
metaheat_sub[["plot"]]

## Generate two metaheatmaps in the same plot. In this exampe one data frame
contains the whole dataset while in the other only reads of 28 nucleotides
are present
sample_name <- "Samp1"
metaheat_df <- list()
metaheat_df[["subsample_28nt"]] <- subset(reads_psite_info[[sample_name]], length == 28)
metaheat_df[["whole_sample"]] <- reads_psite_info[[sample_name]]
names_list <- list("Only_28" = c("subsample_28nt"), "All" = c("whole_sample"))
metaheat_comparison <- metaheatmap_psite(metaheat_df, mm81cdna, sample = names_list)
metaheat_comparison[["plot"]]
}
