% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaprofile_psite}
\alias{metaprofile_psite}
\title{Plot ribosome occupancy metaprofiles at single-nucleotide resolution.}
\usage{
metaprofile_psite(data, annotation, sample, scale_factors = NULL,
  length_range = "all", transcripts = NULL, utr5l = 25, cdsl = 50,
  utr3l = 25, plot_title = "auto")
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{annotation}{A data frame with a reference annotation of the transripts.
It must contain at least four columns named \emph{transcript},
\emph{l_utr5}, \emph{l_cds}, \emph{l_utr3} containing the name of the
transcripts (the same as in the reference transcriptome), the position of
the first nucleotide of the \emph{5' UTR}, the \emph{CDS} and the  \emph{3'
UTR}, respectively. No specific order is required.}

\item{sample}{A character string vector specifying the name of the sample (or
of its replicates) of interest.}

\item{scale_factors}{A numeric vector of scale factors to be used for merging
the replicates of the specified sample (if any). The vector must contain at
least a set of values named after the strings listed in \code{sample}. No
specific order is required.}

\item{length_range}{Either "all", an integer or an integer vector. In the
first case all the reads will be used to generate the metaprofile.
Otherwise, only the reads matching the specified length(s) will be
employed.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the metaprofile. By default this argument is
NULL, which implies all the transcripts in \code{data} will be used. Note
that if either the 5' UTR, the coding sequence or the 3' UTR of a
transcript is shorther than what is specified by \code{utr5l},
\eqn{2*}\code{cdsl} and \code{utr3l} respectively, the transcript will not
be cosidered.}

\item{utr5l}{A positive integer specifying the length (in nucleotides) of the
5' UTR portion that will flank the start codon in the plot. The default
value is 25.}

\item{cdsl}{A positive integer specifying the length (in nucleotides) of the
coding sequence portion that will flank both the start and the stop codon
in the plot. The default value is 50.}

\item{utr3l}{A positive integer specifying the length (in nucleotides) of the
3' UTR portion that will flank the stop codon in the plot. The default
value is 25.}

\item{plot_title}{Either "auto" (the default), NULL or any character string.
When \emph{auto}, the title of the plot is the name of the sample followed
by the number of transcript and the read lengths employed for generating
the metaprofile.}
}
\value{
A list containing a ggplot2 plot object, a data frame with the
  associated data and the transcripts employed for generating the plot.
}
\description{
Builds a metaprofile based on the identified P-site of the reads. It sums up
the number of P-sites per nucleotide for the chosen region, starting from one
ore more replicates of a sample.
}
\examples{
data(reads_psite_list)
data(mm81cdna)

## Generate the metaprofile employing the whole dataset
metaprof_whole <- metaprofile_psite(reads_psite_list, mm81cdna, sample = "Samp1")
metaprof_whole[["plot"]]

## Generate the metaprofile employing reads of 27, 28 and 29 nucleotides and
a specified set of transcripts (for example those with at least one P-site on
the first nucleotide of the coding sequence)
sample_name <- "Samp1"
sub_reads_psite_list <- subset(reads_psite_list[[sample_name]], psite_from_start == 0)
transcript_names <- as.character(sub_reads_psite_list$transcript)
metaprof_sub <- metaprofile_psite(reads_psite_list, mm81cdna, sample = sample_name,
length_range = 27:29, transcripts = transcript_names)
metaprof_sub[["plot"]]
}
