% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaprofile_psite}
\alias{metaprofile_psite}
\title{Ribosome occupancy metaprofiles at single-nucleotide resolution.}
\usage{
metaprofile_psite(data, annotation, sample, scale_factors = NULL,
  length_range = "all", transcripts = NULL, utr5l = 25, cdsl = 50,
  utr3l = 25, plot_title = NULL)
}
\arguments{
\item{data}{List of data tables from \code{\link{psite_info}}.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Either a character string specifying the name of the sample of
interest or a character string vector specifying the name of its
replicates. In the latter case the final metaprofiles are generated by
merging the results for each replicate exploiting the scale factors
specified by \code{scale_factors}.}

\item{scale_factors}{Named numeric vector the same length as \code{sample}
specifying the scale factors for generating metaprofiles from multiple
replicates (see \code{sample}). Please be careful to name each element of
the vector after the correct corresponding string in \code{sample}. No
specific order is required. Default is NULL i.e. all scale factors are
automatically set to 1.}

\item{length_range}{Integer or an integer vector specyfying the read
length(s) to be included in the analysis. Default is "all" i.e. all read
lengths are used.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.
Please note: transcripts with either 5' UTR, coding sequence or 3' UTR
shorter than \code{utr5l}, \eqn{2*}\code{cdsl} and \code{utr3l},
respectively, are automatically discarded.}

\item{utr5l}{Positive integer specifying the length (in nucleotides) of the
5' UTR region flanking the start codon to be considered in the analysis.
Default is 25.}

\item{cdsl}{Positive integer specifying the length (in nucleotides) of the
CDS regions flanking both the start and stop codon to be considered in the
analysis. Default is 50.}

\item{utr3l}{Positive integer specifying the length (in nucleotides) of the
3' UTR region flanking the stop codon to be considered in the analysis.
Default is 25.}

\item{plot_title}{Character string specifying the title of the plot. It can
be any string provided by the user or "sample", "transcript" and
"length_range" for automatically displaying the name of the sample(s)
specified by \code{sample}, the number of transcripts or the read length(s)
employed for generating the metaprofiles, respectively. A combination of
the three strings, dot-separated, can be used for displaying multiple
information. For example, specifying "sample.length_range" the title
reports both the name of the sample(s) and the read length(s). Default is
NULL i.e. no title is displayed.}
}
\value{
A list containing a ggplot2 object ("plot") and the data table with
  the associated data ("dt").
}
\description{
This function generates two metaprofiles displaying the abundance of P-sites
around the start and the stop codon of annotated CDSs.
}
\details{
The intensity of signal in the metaprofiles corresponds, for each
  nucleotide, to the sum of the number of P-sites (defined by their leftmost
  position) mapping on that position for all transcripts in one or multiple
  replicates.
}
\examples{
data(reads_psite_list)
data(mm81cdna)

## Generate metaprofiles employing all read lengths:
metaprof_whole <- metaprofile_psite(reads_psite_list, mm81cdna, sample = "Samp1")
metaprof_whole[["plot"]]

## Generate metaprofiles employing reads of 27, 28 and 29 nucleotides and a
## subset of transcripts (in this example only transcripts with at least one
## P-site mapping on the translation initiation site are kept):
sample_name <- "Samp1"
sub_reads_psite_list <- reads_psite_list[[sample_name]][psite_from_start == 0]
transcript_names <- as.character(sub_reads_psite_list$transcript)
metaprof_sub <- metaprofile_psite(reads_psite_list, mm81cdna, sample = sample_name,
length_range = 27:29, transcripts = transcript_names)
}
