% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{metaprofile_psite}
\alias{metaprofile_psite}
\title{Ribosome occupancy metaprofiles at single-nucleotide resolution.}
\usage{
metaprofile_psite(
  data,
  annotation,
  sample,
  multisamples = "separated",
  plot_style = "split",
  scale_factors = NULL,
  length_range = "all",
  transcripts = NULL,
  frequency = FALSE,
  utr5l = 25,
  cdsl = 50,
  utr3l = 25,
  colour = NULL,
  plot_title = NULL
)
}
\arguments{
\item{data}{List of data tables from \code{\link{psite_info}}.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Either character string or character string vector specifying
the name of the sample(s) of interest. A named list of two or more
character strings and/or character string vectors can be provided. In this
case, each element of the list should include the name of the replicate(s)
of the samples of interest and the name assigned to the elements of the
list are displayed in the plot. Multiple replicates specified in character
string vectors are handled according to \code{multisample}.}

\item{multisamples}{Either "separated", "average" or "sum". It specifies how
to handle multiple samples and replicates. If "saparated", one metaprofile
for each sample included in \code{sample} is returned as an independent
ggplot object. If "average" or "sum" i) one metaprofiles is returned if
\code{sample} is a character string vector or ii) one metaprofiles is built
for each element of \code{sample} when it is a list. If "average", the
metaprofiles display for each nucleotide the mean signal and the
corresponding standard error computed among the replicates. If "sum",
the metaprofiles display for each nucleotide the sum of the signal of the
replicates. In both cases a single ggplot object is returned, where
multiple metaprofiles are organized and displayed according to
\code{plot_style}. Default is "separated".}

\item{plot_style}{Either "split", "ovelaid" or "mirrored". It specifies how
to organize and display multiple metaprofiles. If "split", the metaprofiles
are placed one below the other, in independent boxes. If "overlaid", all
metaprofiles are superimposed. If "mirrored" \code{sample} must be a list
of exactly two elements and the two metaprofiles are mirrored along the x
axis. Default is "split".}

\item{scale_factors}{Named numeric vector the same length as \code{sample}
specifying the scale factors for generating metaprofiles from multiple
replicates (see \code{sample}). Please be careful to name each element of
the vector after the correct corresponding string in \code{sample}. No
specific order is required. When \code{frequency} is TRUE, this parameter
is not considered. Default is NULL i.e. all scale factors are automatically
set to 1.}

\item{length_range}{Integer or integer vector specyfying the read
length(s) to be included in the analysis. Default is "all" i.e. all read
lengths are used.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.
Please note: transcripts with either 5' UTR, coding sequence or 3' UTR
shorter than \code{utr5l}, \eqn{2*}\code{cdsl} and \code{utr3l},
respectively, are automatically discarded.}

\item{frequency}{Logical value whether to normalize the metaprofile(s) such
that the area under the curve(s) is 1. If TRUE and \code{multisamples} is
set to "average" or "sum", the normalization is performed before combining
the signal from multiple samples. Default is FALSE.}

\item{utr5l}{Positive integer specifying the length (in nucleotides) of the
5' UTR region flanking the start codon to be considered in the analysis.
Default is 25.}

\item{cdsl}{Positive integer specifying the length (in nucleotides) of the
CDS regions flanking both the start and stop codon to be considered in the
analysis. Default is 50.}

\item{utr3l}{Positive integer specifying the length (in nucleotides) of the
3' UTR region flanking the stop codon to be considered in the analysis.
Default is 25.}

\item{colour}{Character string or character string vector specifying the
colour of the metaprofile(s). If \code{sample} is a list of multiple
elements and \code{multisamples} is set to "average" or "sum", a colour for
each element of the list is required. If this parameter is not specified
the R default palette is employed. Default is NULL.}

\item{plot_title}{Character string specifying the title of the plot. It can
be any string provided by the user or "sample", "transcript" and
"length_range" for automatically displaying the name of the sample(s)
specified by \code{sample}, the number of transcripts and the most frequent
read lengths (i.e. associated to 90% of the signal) employed for generating
the metaprofiles, respectively. A combination of the three strings,
dot-separated, can be used for displaying multiple information. For
example, specifying "sample.length_range" the title reports both the name
of the sample(s) and the read length(s). Default is NULL i.e. no title is
displayed.}
}
\value{
A list containing one or more ggplot2 object(s) and the data table
  with the associated data ("dt").
}
\description{
This function generates metaprofiles displaying the abundance of P-sites
around the start and the stop codon of annotated CDSs. For one sample the
intensity of the signal in the metaprofiles corresponds, for each nucleotide,
to the sum of the number of P-sites (defined by their leftmost position)
mapping on that position for all transcripts. Multiple samples can be handled
in several ways.
}
\examples{
## data(reads_list)
## data(mm81cdna)
##
## ## compute and add p-site datails
## psite_offset <- psite(reads_list, flanking = 6, extremity = "auto")
## reads_psite_list <- psite_info(reads_list, psite_offset)
##
## ## Generate metaprofiles employing all read lengths:
## metaprof_whole <- metaprofile_psite(reads_psite_list, mm81cdna,
##                                     sample = "Samp1")
## metaprof_whole[["plot_Samp1"]]
## 
## ## Generate metaprofiles employing reads of 27, 28 and 29 nucleotides and
## ## a subset of transcripts (in this example only transcripts with at least
## ## one P-site mapping on the translation initiation site are kept):
## sample_name <- "Samp1"
## sub_reads_psite_list <- reads_psite_list[[sample_name]][psite_from_start == 0]
## transcript_names <- as.character(sub_reads_psite_list$transcript)
## metaprof_sub <- metaprofile_psite(reads_psite_list, mm81cdna,
##                                   sample = sample_name,
##                                   length_range = 27:29,
##                                   transcripts = transcript_names)
}
