% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm81cdna}
\alias{mm81cdna}
\title{Annotation data table}
\format{
A data table with 25,892 rows and 5 variables:
\describe{
  \item{transcript}{Name of the transcript (ENST ID and version, dot
  separated)}
  \item{l_tr}{Length of the transcript, in nucleotides}
  \item{l_utr5}{Length of the annotated 5' UTR (if any), in nucleotides}
  \item{l_cds}{Length of the annotated CDS (if any), in nucleotides}
  \item{l_utr3}{Length of the annotated 3' UTR (if any), in nucleotides}
}
}
\usage{
mm81cdna
}
\description{
A dataset containing basic information about 25,892 mouse mRNAs (Ensembl v81
transcript annotation).
}
\keyword{datasets}
