% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psites.R
\name{psite}
\alias{psite}
\title{Identify the ribosome P-site position within the reads.}
\usage{
psite(data, flanking = 6, start = TRUE, extremity = "auto",
  plot = FALSE, plotdir = NULL, plotformat = "png", cl = 99)
}
\arguments{
\item{data}{A list of data tables from either \code{\link{bamtolist}} or
\code{\link{bedtolist}}.}

\item{flanking}{An integer that specifies how many nucleotides, at least, of
the reads mapping on the reference codon must flank the reference codon in
both directions. Default is 6.}

\item{start}{A logical value whether ot not to compute the P-site offsets
starting from the reads aligning on the translation initiation site. FALSE
implies that the reads mapping on the last triplet before the stop codon
are used instead. Default is TRUE.}

\item{extremity}{A character string specifing which extremity of the reads
should be used in the correction step of the algorithm. It can be either
"5end" or "3end" for the 5' and the 3' extremity, respectively. Default is
"auto", meaning that the best extremity is automatically selected.}

\item{plot}{A logical value whether or not to plot the occupancy metaprofiles
showing the P-sites offsets computed throughout the two steps of the
algorithm. Default is FALSE.}

\item{plotdir}{A character string specifying the (existing or not) location
of the directory where the occupancy metaprofiles shuold be stored. This
parameter is considered only if \code{plot} is TRUE. By default this
argument is NULL, which implies it is set as a subfolder of the working
directory, called \emph{offset_plot}.}

\item{plotformat}{Either "png" (the default) or "pdf", this parameter
specifies the file format of the generated metaprofiles. It is considered
only if \code{plot} is TRUE.}

\item{cl}{An integer value in \emph{[1,100]} specifying the confidence level
for restricting the generation of the occupancy metaprofiles to a sub-range
of read lengths. By default it is set to 99. This parameter is considered
only if \code{plot} is TRUE.}
}
\value{
A data table.
}
\description{
This function identifies within each read the position of the ribosome
P-site, determined by the localisation of its first nucleotide. The function
processes the samples separately starting from the reads aligning on the
reference codon (selected by the user between the start codon and the second
to last codon) of any annotated coding sequence. It then returns the position
of the P-site specifically inferred for all the read lengths. It also allows
to plot a collection of read length-specific occupancy metaprofiles
showing the P-sites offsets computed throughout the two steps of the
algorithm.
}
\examples{
data(reads_list)

## Compute the P-site offset automatically selecting the otimal read
## extremity for the correction step and not plotting any metaprofile
psite(reads_list, flanking = 6, extremity="auto")

## Compute the P-site offset specifying the extremity used in the correction
## step and plotting the metaprofiles only for a sub-range of read lengths (the
## middle 95\%). The plots will be placed in the current working directory.
psite_offset <- psite(reads_list, flanking = 6, extremity="3end", plot = TRUE, cl = 95)
}
