% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psites.R
\name{psite_info}
\alias{psite_info}
\title{Updates reads information adding features associated to the inferred P-sites.}
\usage{
psite_info(data, offset, sequence = NULL)
}
\arguments{
\item{data}{A list of data frames from \code{\link{bamtolist}}}

\item{offset}{A data frame from \code{\link{psite}}.}

\item{sequence}{A list of character strings containing the nucleotide
sequence of the transcripts. Both the sequences and their names (i.e. the
name of the associated transcripts used to identify the elements of the
list) must be the same as in the reference transcriptome. Use this argument
to attach to the data frames an additional column containing the three
nucletotides corresponding to the identified P-sites. Default is NULL.}
}
\value{
A list of data frames.
}
\description{
Ipdates the data frames of the input list adding information associated to
the P-site positions identfied by \code{\link{psite}}. It attaches to the
data frames 4 columns containing i) the offset between the P-site within the
reads and both the start and the stop codon of the corresponding transcript;
ii) the region of the transcript (5' UTR, CDS, 3' UTR) that includes the
P-site; iii) the sequence of the triplet covered by the P-site.
}
\examples{
data(reads_list)
data(psite_offset)
data(mm81cdna)

reads_psite_list <- psite_info(reads_list, psite_offset, seq = NULL)
}

