% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psites.R
\name{psite_info}
\alias{psite_info}
\title{Update reads information according to the inferred P-sites.}
\usage{
psite_info(data, offset, fastapath = NULL, fasta_genome = TRUE,
  bsgenome = NULL, gtfpath = NULL, txdb = NULL, dataSource = NA,
  organism = NA, granges = FALSE)
}
\arguments{
\item{data}{A list of data frames from either \code{\link{bamtolist}} or
\code{\link{bedtolist}}.}

\item{offset}{A data frame from \code{\link{psite}}.}

\item{fastapath}{An optional character string specifying the path to the
FASTA file used in the alignment step, including its name and extension.
This file can contain reference nucleotide sequences either of a genome
assembly or of all the transcripts (see \code{fasta_genome}). Please make
sure the sequences derive from the same release of the annotation file used
in the \code{\link{create_annotation}} function. Note: either
\code{fastapath} or \code{bsgenome} is required to generate an
additional column reporting the three nucletotides covered by the P-sites.
Default is NULL.}

\item{fasta_genome}{A locigal value whether or not the FASTA file specified
by \code{fastapath} contains nucleotide sequences of a genome assembly.
FALSE means that the nucleotide sequences of all the transcripts are
provided instead. When this parameter is TRUE (the default), an annotation
object is required (see \code{gtfpath} and \code{txdb}).}

\item{bsgenome}{An optional character string specifying the name of the
BSgenome data package containing the genome sequences to be loaded. If it
is not already present in your system, it will be installed through the
biocLite.R script (check the list of data packages available in the
Bioconductor repositories for your version of R/Bioconductor by the
\code{\link[BSgenome]{available.genomes}} function of the BSgenome
package)). This parameter also requires an annotation object (see
\code{gtfpath} and \code{txdb}). Please make sure the sequences included in
the specified BSgenome data pakage are in agreement with the sequences used
in the alignment step. Note: either \code{fastapath} or \code{bsgenome}
is required to generate an additional column reporting the three
nucletotides covered by the P-sites. Default is NULL.}

\item{gtfpath}{A character string specifying the path to te GTF file,
including its name and extension. Please make sure the GTF derives from the
same release of what is specified by \code{fastapath} or by
\code{bsgenome}. Note that either \code{gtfpath} or \code{txdb} must be
specified when the nucleotide sequences of a genome assembly are provided
(see \code{fastapath} or \code{bsgenome}). Default is NULL.}

\item{txdb}{A character string specifying the name of the annotation package
for TxDb object(s) to be loaded. If it is not already present in your
system, it will be installed through the biocLite.R script (check the list
of TxDb annotation packages available in the Bioconductor repositories at
http://bioconductor.org/packages/release/BiocViews.html#___TxDb )). Please
make sure the annotation package derives from the same release of what is
specified by \code{fastapath} or by \code{bsgenome}. Note that either
\code{gtfpath} or \code{txdb} must be specified when the nucleotide
sequences of a genome assembly are provided (see \code{fastapath} or
\code{bsgenome}). Default is NULL.}

\item{dataSource}{An optional character string describing the origin of the
GTF data file. For more information about this parameter please refer to
the description of \emph{dataSource} of the
\code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in the
\code{GenomicFeatures} package.}

\item{organism}{A optional character string reporting the genus and species
of the organism when \code{gtfpath} is specified.  For more information
about this parameter please refer to the description of \emph{dataSource}
of the \code{\link[GenomicFeatures]{makeTxDbFromGFF}} function included in
the \code{GenomicFeatures} package.}

\item{granges}{A logical value whether or not to return a GRangesList object.
Default is FALSE, meaning that a list of data frames (the required input
for the downstream analyses and graphical outputs provided by riboWaltz) is
returned instead.}
}
\value{
A list of data frames or a GRangesList object.
}
\description{
Starting ftom the P-site position identfied by \code{\link{psite}}, this
function updates the data frames that contains information about the reads.
It attaches to the data frames 4 columns reporting the P-site position with
respect to the 1st nucleotide of the transcript, the start and the stop codon
of the annotated coding sequence (if any) and the region of the transcript
(5' UTR, CDS, 3' UTR) that includes the P-site. Please note: if a transcript
is not associated to any annotated CDS then the positions of the P-site from
both the start and the stop codon is set to NA. If either a FASTA file or a
BSgenome data package with the nucleotide  sequences is provided, an
additional column reporting the three nucleotides covered by the P-site is
attached.
}
\examples{
data(reads_list)
data(psite_offset)
data(mm81cdna)

reads_psite_list <- psite_info(reads_list, psite_offset)
}
