% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psites.R
\name{psite_info}
\alias{psite_info}
\title{Update reads information according to the inferred P-sites.}
\usage{
psite_info(data, offset, granges = FALSE, fastapath = NULL,
  bsgenome_dp = NULL, txdb = NULL)
}
\arguments{
\item{data}{A list of data frames from either \code{\link{bamtolist}} or
\code{\link{bedtolist}}.}

\item{offset}{A data frame from \code{\link{psite}}.}

\item{granges}{A logical value whether or not to return a GRangesList 
object. Default is FALSE, meaning that a list of data frames (the required
input for the downstream analyses and graphical outputs provided by
riboWaltz) is returned instead.}

\item{fastapath}{A character string specifying the path to the FASTA file
containing the reference nucleotide sequences. Please make sure that the
sequences and their names derive from the same release of and are in
agreement with the annotation file used in the
\code{\link{create_annotation}} function. Either \code{fastapath} or
\code{bsgenome_dp} coulped with \code{txdb} must be specified to attach an
additional column reporting the three nucletotides covered by the
identified P-sites. Default is NULL.}

\item{bsgenome_dp}{A character string specifying the name of the BSgenome
data package to be loaded. If the specified data package is not already
present in your system, it is installed through the biocLite.R script.
Please check the data packages available in the Bioconductor repositories
for your version of R/Bioconductor using the
\code{\link[BSgenome]{available.genomes}} function from the BSgenome package.
Either \code{fastapath} or \code{bsgenome_dp} coulped with \code{txdb} must
be specified to attach an additional column reporting the three
nucletotides covered by the identified P-sites. Default is NULL.}

\item{txdb}{A TxDb object. This parameter is considered only if
\code{bsgenome_dp} is specified. Default is NULL.}
}
\value{
A list of data frames or a GRangesList object.
}
\description{
Starting ftom the P-site position identfied by \code{\link{psite}}, this
function updates the data frames that contains information about the reads.
It attaches to the data frames 4 columns reporting the P-site position with
respect to the 1st nucleotide of the transcript, the start and the stop codon
of the annotated coding sequence (if any) and the region of the transcript
(5' UTR, CDS, 3' UTR) that includes the P-site. Please note: if a transcript
is not associated to any annotated CDS then the positions of the P-site from
both the start and the stop codon is set to NA. If either a FASTA file or a
BSgenome data package with the nucleotide  sequences is provided, an
additional column reporting the three nucleotides covered by the P-site is
attached.
}
\examples{
data(reads_list)
data(psite_offset)
data(mm81cdna)

reads_psite_list <- psite_info(reads_list, psite_offset)
}
