% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psite_offset}
\alias{psite_offset}
\title{P-site offsets}
\format{A data table with 31 rows and 9 variables (the lengths and the
  distances are expressed in nucleotides):
\describe{
  \item{length}{Length of the read}
  \item{total_percentage}{Percentage of reads of the considered length in the
  whole dataset}
  \item{start_percentage}{Percentage of reads of the considered length
  aligning on the start codon (if any)}
  \item{around_start}{A logical value reporting whether at least one read of
  the specified length aligns on the start codon (T = yes, F = no)}
  \item{offset_from_5}{Temporary P-site offset from the 5' end of read
  (before the correction step)}
  \item{offset_from_3}{Temporary P-site offset from the 3' end of read
  (before the correction step)}
  \item{adj_offset_from_5}{P-site offset from the 5' end of read after the
  correction step}
  \item{adj_offset_from_3}{P-site offset from the 3' end of read after the
  correction step}
  \item{sample}{Name of the sample}
}}
\usage{
psite_offset
}
\description{
This dataset contains information on the offset computed by \code{\link{psite}}
starting from \code{\link{reads_list}}.
}
\keyword{datasets}
