% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_psite_list}
\alias{reads_psite_list}
\title{Reads information data table updated with P-site details}
\format{A list of data tables with 1 object (named \emph{Samp1}) of 393,338
  rows and 10 variables:
\describe{
  \item{transcript}{Name of the transcript (ENST ID and version, dot
  separated)}
  \item{end5}{Position of the 5' end of the read with respect to the first
  nucleotide of the transcript, in nucleotides}
  \item{psite}{Position of the P-site with respect to the first nucleotide of
  the transcript, in nucleotides}
  \item{end3}{Position of the 3' end of the read with respect to the first
  nucleotide of the transcript, in nucleotides}
  \item{length}{Length of the read, in nucleotides (intended as the width of
  the reference sequence region covered by the RNA fragment. For further
  information about this choice please refer to section \code{Details} of
  function \code{\link{bamtolist}})}
  \item{cds_start}{Leftmost position of the CDS with respect to the first
  nucleotide of the transcript, in nucleotides}
  \item{cds_stop}{Rightmost position of the CDS with respect to the first
  nucleotide of the transcript, in nucleotides}
  \item{psite_from_start}{Position of the P-site with respect to the first
  nucleotide of the annotated CDS (if any), in nucleotides}
  \item{psite_from_stop}{Position of the P-site with respect to the last
  nucleotide of the annotated CDS (if any), in nucleotides}
  \item{psite_region}{Region of the transcript that includes the P-site
  (5utr, cds, 3utr)}
}}
\usage{
reads_psite_list
}
\description{
An example dataset that combines details on reads mapping on the mouse
transcriptome (see \code{\link{reads_list}}) and length-specific ribosome
P-site offsets (see \code{\link{psite_offset}}), as returned by
\code{\link{psite_info}}.
}
\keyword{datasets}
