% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_regions.R
\name{region_psite}
\alias{region_psite}
\title{Plot percentage of P-site per transcript region.}
\usage{
region_psite(data, annotation, sample = NULL, transcripts = NULL,
  label = NULL, colour = c("gray70", "gray40", "gray10"))
}
\arguments{
\item{data}{A list of data frames from \code{\link{psite_info}}.}

\item{annotation}{A data frame from \code{\link{create_annotation}}.}

\item{sample}{A character string vector specifying the name of the sample(s)
of interest. By default this argument is NULL, meaning that all the data
frames in \code{data} will be analysed.}

\item{transcripts}{A character string vector specifying the name of the 
transcripts to be considered in the analysis. By default this argument is 
NULL, which implies all the transcripts in \code{data} will be used.
Transcripts with no annotated \emph{5' UTR}, \emph{CDS} and \emph{3'UTR}
will be removed.}

\item{label}{A character string vector of the same length of \code{sample}
specifying the name of the samples to be displaied in the plot. By default
this argument is NULL i.e. the labels correspond to the sample names.}

\item{colour}{A character string vector of three elements specifying the
colours of the bars (for \emph{5' UTR}, \emph{CDS} and \emph{3'UTR}
respectively). By default is a grayscale.}
}
\value{
A list containing a ggplot2 object, and a data frame with the
  associated data.
}
\description{
For each sample computes the percentage of P-sites falling in the three 
regions of the transcripts with annotated 5' UTR, CDS and 3'UTR and generates
the corresponding barplot. The function also calculates the percentage of
region length for the selected transcripts (reported in column "RNAs").
}
\examples{
data(reads_psite_list)
data(mm81cdna)

reg_psite <- region_psite(reads_psite_list, mm81cdna, sample = "Samp1")
reg_psite[["plot"]]
}

