% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_regions.R
\name{region_psite}
\alias{region_psite}
\title{Percentage of P-sites per transcript region.}
\usage{
region_psite(
  data,
  annotation,
  sample = NULL,
  transcripts = NULL,
  label_sample = NULL,
  colour = c("gray70", "gray40", "gray10")
)
}
\arguments{
\item{data}{Either list of data tables or GRangesList object from
\code{\link{psite_info}}.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Character string vector specifying the name of the sample(s) of
interest. Default is NULL i.e. all samples in \code{data} are processed.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.
Please note: transcripts without annotated 5' UTR, CDS and 3' UTR are
automatically discarded.}

\item{label_sample}{Named character string vector the same length as
\code{sample} specifying the sample names to be displayed in the plot.
Plase be careful to name each element of the vector after the correct
corresponding string in \code{sample}. No specific order is required.
Default is NULL i.e. sample names in \code{sample} are used.}

\item{colour}{Character string vector of three elements specifying the colour
for the 5' UTR, CDS and 3' UTR bars, respectively. Default is a grayscale.}
}
\value{
A list containing a ggplot object ("plot") and the data table with
  the associated data ("dt").
}
\description{
This function computes the percentage of P-sites falling in the three
annotated regions of the transcripts (5' UTR, CDS and 3' UTR) and generates a
bar plot of the resulting values.
}
\details{
Column "RNAs" reports the percentage of region length for the
  transcripts included in the analysis, based on the cumulative nucleotide
  length of 5' UTRs, CDSs and 3' UTRs. These values reflect the expected read
  distribution from a random fragmentation of RNA and can be used as a
  baseline to verify the expected enrichment of ribosome (P-site) signal in
  CDSs.
}
\examples{
## data(reads_list)
## data(mm81cdna)
##
## ## compute and add p-site datails
## psite_offset <- psite(reads_list, flanking = 6, extremity = "auto")
## reads_psite_list <- psite_info(reads_list, psite_offset)
##
## reg_psite <- region_psite(reads_psite_list, mm81cdna, sample = "Samp1")
## reg_psite[["plot"]]
}
