% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_end_metaheatmap_plot.R
\name{rends_heat}
\alias{rends_heat}
\title{Metaheatmaps of the two extremities of the reads.}
\usage{
rends_heat(
  data,
  annotation,
  sample,
  transcripts = NULL,
  cl = 95,
  utr5l = 50,
  cdsl = 50,
  utr3l = 50,
  log_colour = F,
  colour = "black"
)
}
\arguments{
\item{data}{Either list of data tables or GRangesList object from
\code{\link{bamtolist}}, \code{\link{bedtolist}},
\code{\link{length_filter}} or \code{\link{psite_info}}.}

\item{annotation}{Data table as generated by \code{\link{create_annotation}}.}

\item{sample}{Character string specifying the name of the sample of interest.}

\item{transcripts}{Character string vector listing the name of transcripts to
be included in the analysis. Default is NULL i.e. all transcripts are used.
Please note: transcripts with either 5' UTR, coding sequence or 3' UTR
shorter than \code{utr5l}, \eqn{2*}\code{cdsl} and \code{utr3l},
respectively, are automatically discarded.}

\item{cl}{Integer value in [1,100] specifying a confidence level for
restricting the plot to a sub-range of read lengths. The new range is
associated to the most abundant populations of reads accounting for the cl%
of the sample. Default is 95.}

\item{utr5l}{Positive integer specifying the length (in nucleotides) of the
5' UTR region flanking the start codon to be considered in the analysis.
Default is 50.}

\item{cdsl}{Positive integer specifying the length (in nucleotides) of the
CDS regions flanking both the start and stop codon to be considered in the
analysis. Default is 50.}

\item{utr3l}{Positive integer specifying the length (in nucleotides) of the
3' UTR region flanking the stop codon to be considered in the analysis.
Default is 50.}

\item{log_colour}{Logical value whether to use a logarithmic colour scale
(strongly suggested in case of large signal variations). Default is FALSE.}

\item{colour}{Character string specifying the colour of the plot. The colour
scheme is as follow: tiles corresponding to the lowest signal are always
white, tiles corresponding to the highest signal are of the specified
colour and the progression between these two colours follows either linear
or logarithmic gradients (see \code{log_colour}). Default is "black".}
}
\value{
A list containing a ggplot2 object ("plot") and the data table with
  the associated data ("dt").
}
\description{
This function generates four metaheatmaps displaying the abundance of the 5'
and 3' extremity of reads mapping around the start and the stop codon of
annotated CDSs, stratified by their length.
}
\examples{
data(reads_list)
data(mm81cdna)

## Generate metaheatmaps for all read lengths:
heatend_whole <- rends_heat(reads_list, mm81cdna, sample = "Samp1", cl = 100)

## Generate metaheatmaps for a sub-range of read lengths shortening the
## flanking regions around the start and stop codon:
heatend_sub95 <- rends_heat(reads_list, mm81cdna, sample = "Samp1", cl = 95,
utr5l = 30, cdsl = 40, utr3l = 30)
}
