% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_end_metaheatmap_plot.R
\name{rends_heat}
\alias{rends_heat}
\title{Plot metaheatmaps based on the two extremities of the reads.}
\usage{
rends_heat(data, annotation, sample, transcripts = NULL, cl = 95,
  utr5l = 50, cdsl = 50, utr3l = 50, log = F, colour = "black")
}
\arguments{
\item{data}{A list of data tables from either \code{\link{bamtolist}} or
\code{\link{bedtolist}}.}

\item{annotation}{A data table as generated by
\code{\link{create_annotation}}.}

\item{sample}{A character string specifying the name of the sample of
interest.}

\item{transcripts}{A character string vector specifying the name of the
transcripts to be included in the analysis. By default this argument is
NULL, meaning that all the transcripts in \code{data} are used. Note that
if either the 5' UTR, the coding sequence or the 3' UTR of a transcript is
shorther than \code{utr5l}, \eqn{2*}\code{cdsl} and \code{utr3l}
respectively, the transcript is automatically discarded.}

\item{cl}{An integer value in \emph{[1,100]} specifying the confidence level
for restricting the plot to a sub-range of read lengths. Default is 95.}

\item{utr5l}{A positive integer specifying the length (in nucleotides) of the
5' UTR region that in the plot flanks the start codon. The default value is
50.}

\item{cdsl}{A positive integer specifying the length (in nucleotides) of the
CDS region that in the plot will flank both the start and the stop codon.
The default value is 50.}

\item{utr3l}{A positive integer specifying the length (in nucleotides) of the
3' UTR region that in the plot flanks the start codon. The default value is
50.}

\item{log}{A logical value whether or not to use a logarithmic scale colour
(strongly suggested in case of large variations of the signal). Default is
FALSE.}

\item{colour}{A character string specifying the colour of the plot. Default
is "black".}
}
\value{
A list containing a ggplot2 object, and a data table with the
  associated data.
}
\description{
For a specified sample this function plots four metaheatmaps showing the
abundance of the 5' and the 3' end of the reads mapping around the start and
the stop codon of the annotated CDS (if any), stratified by their length. It
is possible to visualise the metaheatmaps for all the read lengths or to
restrict the graphical output to a sub-range of read lengths.
}
\examples{
data(reads_list)
data(mm81cdna)

## Visualise the metaheatmaps for all the read lengths
heatend_whole <- rends_heat(reads_list, mm81cdna, sample = "Samp1", cl = 100)

## Visualise the metaheatmaps for a sub-range of read lengths (the middle
## 95\%) reducing the flanking regions around the start and the stop codon
heatend_sub95 <- rends_heat(reads_list, mm81cdna, sample = "Samp1", cl = 95,
utr5l = 30, cdsl = 40, utr3l = 30)
}
