% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_plot.R
\name{rlength_distr}
\alias{rlength_distr}
\title{Plot read length distributions.}
\usage{
rlength_distr(data, sample, cl = 100)
}
\arguments{
\item{data}{A list of data frames from \code{\link{bamtolist}}.}

\item{sample}{A character string specifying the name of the sample of
interest.}

\item{cl}{An integer with value in \emph{[1,100]} specifying the read length
confidence level for restricting the distribution to a chosen range of
lengths. By default it is set to 100, i.e. the whole set of reads lengths
will be considered.}
}
\value{
A list containing a ggplot2 plot object, and a data frame with the
  associated data.
}
\description{
Plots the read length distribution for a specified sample of the input list.
It is possible to look at the length distribution based on the whole set of
reads or to restrict it to a sub-range of read lengths.
}
\examples{
data(reads_list)

## Visualise the whole read length distribution
lendist_whole <- rlength_distr(reads_list, sample = "Samp1", cl = 100)
lendist_whole[["plot"]]

## Visualise the middle 95\% of the read length distribution
lendist_sub95 <- rlength_distr(reads_list, sample = "Samp1", cl = 95)
lendist_sub95[["plot"]]
}

