% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_plot.R
\name{rlength_distr}
\alias{rlength_distr}
\title{Plot read length distributions.}
\usage{
rlength_distr(data, sample, cl = 100)
}
\arguments{
\item{data}{A list of data tables from either \code{\link{bamtolist}} or
\code{\link{bedtolist}}.}

\item{sample}{A character string specifying the name of the sample of
interest.}

\item{cl}{An integer value in \emph{[1,100]} specifying the confidence level
for restricting the plot to a sub-range of read lengths. By default it is
set to 100, meaning that the whole distribution is displayed.}
}
\value{
A list containing a ggplot2 object, and a data table with the
  associated data.
}
\description{
For a specified sample this function plots the read length distribution. It
is possible to visualise the distribution for all the read lengths or to
restrict the graphical output to a sub-range of read lengths.
}
\examples{
data(reads_list)

## Visualise distribution for all the read lengths
lendist_whole <- rlength_distr(reads_list, sample = "Samp1", cl = 100)
lendist_whole[["plot"]]

## Visualise the metaheatmaps for a sub-range of read lengths (the middle 95\%)
lendist_sub95 <- rlength_distr(reads_list, sample = "Samp1", cl = 95)
lendist_sub95[["plot"]]
}
